src/main/java/edu/depaul/cdm/se452/fall2023group1/books/BookController.java

package edu.depaul.cdm.se452.fall2023group1.books;

import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;
import java.util.List;
import java.util.Optional;
import org.springframework.http.ResponseEntity;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.RestController;
@RestController
@RequestMapping("/api/books")
public class BookController {
    @Autowired
    private BookService service;

    @GetMapping
    public List<Book> list() {
        return service.list();
    }

    @PostMapping
    public Book add(@Valid @RequestBody Book book) {
        return service.add(book);
    }

    @PutMapping("/{id}")
    public Book update(@PathVariable Long id, @RequestBody Book book) {
        return service.update(id, book);
    }

    @DeleteMapping("/{id}")
    public void delete(@PathVariable Long id) {
        service.delete(id);
    }

    @GetMapping("/{id}")
    public Optional<Book> findById(@PathVariable Long id) {
        return service.findById(id);
    }


    @ExceptionHandler(BookNotFoundException.class)
    public ResponseEntity<?> handleExceptions(BookNotFoundException ex) {
        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
    }
}
